주문 기능 추가 / 테스트 코드 추가

+++ src/main/java/com/hhplush/eCommerce/api/order/dto/response/ResponseCreateOrderDTO.java
@@ -1,13 +1,13 @@
 package com.hhplush.eCommerce.api.order.dto.response;
 
-import com.hhplush.eCommerce.domain.enums.OrderState;
+import com.hhplush.eCommerce.domain.order.OrderState;
 import java.time.LocalDateTime;
 import lombok.Builder;
 
 @Builder
 public record ResponseCreateOrderDTO(
-    Integer orderId,
-    Integer userId,
+    Long orderId,
+    Long userId,
     OrderState orderStatus,
     LocalDateTime orderAt
 ) {
